+++ xmake/templates/pascal/console/project/xmake.lua
@@ -1,0 +1,7 @@
+add_rules("mode.debug", "mode.release")
+
+target("${TARGETNAME}")
+    set_kind("binary")
+    add_files("src/*.pas")
+
+${FAQ}